Create demo class
